+++ src/wrapper.cpp
@@ -28,6 +28,16 @@ List from_packed_address_wrapper(List x) {
 }
 
 // [[Rcpp::export]]
+CharacterVector to_binary_address_wrapper(List x) {
+  return IpAddressVector(x).encodeBinary();
+}
+
+// [[Rcpp::export]]
+List from_binary_address_wrapper(CharacterVector x) {
+  return IpAddressVector::decodeBinary(x).encodeR();
+}
+
+// [[Rcpp::export]]
 List parse_network_wrapper(CharacterVector x, bool strict) {
   return IpNetworkVector(x, strict).encodeR();
 }